+++ lib/src/models/video_with_comments.dart
@@ -130,8 +130,9 @@ class VideoWithComments extends Video {
       channelId: map['channel_id'],
       channel: ChannelMin.fromMap(map['channel']),
       language: map['lang'],
-      comments:
-          List<Comment>.from(map['comments']?.map((x) => Comment.fromMap(x))),
+      comments: List<Comment>.from(
+        map['comments']?.map((x) => Comment.fromMap(x)) ?? [],
+      ),
     );
   }
 
@@ -145,24 +146,12 @@ class VideoWithComments extends Video {
   bool get stringify => true;
 
   @override
-  List<Object> get props => [
-        id,
-        title,
-        type,
-        topicId ?? 'Topic id not given',
-        publishedAt ?? 'Published at not given',
-        availableAt,
-        duration ?? 'Duration not provided',
-        status,
-        startScheduled ?? 'Scheduled start not given',
-        startActual ?? 'Actual start not given',
-        endActual ?? 'Actual end not given',
-        liveViewers ?? 'Live viewers not given',
-        description ?? 'Description not given',
-        songcount ?? 'Song count not given',
-        language ?? 'Language not given',
-        channelId ?? 'Channel id not given',
-        channel ?? 'Channel not given',
-        comments
-      ];
+  List<Object> get props {
+    final videoProps = super.props;
+    videoProps.addAll([
+      'comments: $comments',
+    ]);
+    
+    return videoProps;
+  }
 }