+++ lib/src/models/video_with_comments_list.dart
@@ -45,7 +45,11 @@ class VideoWithCommentsList extends Equatable {
       total: total,
       paginated: map['paginated'] ?? false,
       videos: List<VideoWithComments>.from(
-          map['items']?.map((x) => VideoWithComments.fromMap(x))),
+        map['items']?.map(
+              (x) => VideoWithComments.fromMap(x),
+            ) ??
+            [],
+      ),
     );
   }
 
@@ -58,5 +62,11 @@ class VideoWithCommentsList extends Equatable {
   bool get stringify => true;
 
   @override
-  List<Object> get props => [total ?? '', paginated, videos];
+  List<Object> get props {
+    return [
+      'total: $total',
+      'paginated: $paginated',
+      'videos: $videos',
+    ];
+  }
 }